bugfix: Legion Transform Fix (#5173)
Legion Transform Fix

+++ code/modules/mob/living/simple_animal/hostile/megafauna/legion.dm
@@ -37,7 +37,9 @@ Difficulty: Medium
 	retreat_distance = 5
 	minimum_distance = 5
 	pixel_x = -32
+	base_pixel_x = -32
 	pixel_y = -16
+	base_pixel_y = -16
 	maptext_height = 96
 	maptext_width = 96
 	ranged_cooldown_time = 20
@@ -56,10 +58,9 @@ Difficulty: Medium
 	stat_attack = UNCONSCIOUS // Overriden from /tg/ - otherwise Legion starts chasing its minions
 
 
-
 /mob/living/simple_animal/hostile/megafauna/legion/Initialize(mapload)
 	. = ..()
-	transform *= 2
+	update_transform(2)
 	ADD_TRAIT(src, TRAIT_NO_FLOATING_ANIM, INNATE_TRAIT)
 	AddElement(/datum/element/simple_flying)
 
@@ -67,16 +68,17 @@ Difficulty: Medium
 /mob/living/simple_animal/hostile/megafauna/legion/enrage()
 	health = 1250
 	maxHealth = 1250
-	transform /= 1.5
+	update_transform(0.66)
 	loot = list(/datum/nothing)
 	crusher_loot = list(/datum/nothing)
-	var/mob/living/simple_animal/hostile/megafauna/legion/legiontwo = new /mob/living/simple_animal/hostile/megafauna/legion(get_turf(src))
-	legiontwo.transform /= 1.5
+	var/mob/living/simple_animal/hostile/megafauna/legion/legiontwo = new(get_turf(src))
+	legiontwo.update_transform(0.66)
 	legiontwo.loot = list(/datum/nothing)
 	legiontwo.crusher_loot = list(/datum/nothing)
 	legiontwo.health = 1250
 	legiontwo.maxHealth = 1250
 	legiontwo.enraged = TRUE
+
 
 /mob/living/simple_animal/hostile/megafauna/legion/unrage()
 	. = ..()
@@ -86,7 +88,9 @@ Difficulty: Medium
 			other.crusher_loot = list(/obj/item/storm_staff, /obj/item/crusher_trophy/empowered_legion_skull)
 			other.maxHealth = 2500
 			other.health = 2500
-	qdel(src) //Suprise, it's the one on lavaland that regrows to full.
+	if(!QDELETED(src))
+		qdel(src) //Suprise, it's the one on lavaland that regrows to full.
+
 
 /mob/living/simple_animal/hostile/megafauna/legion/death(gibbed)
 	for(var/mob/living/simple_animal/hostile/megafauna/legion/other in GLOB.mob_list)
@@ -95,15 +99,8 @@ Difficulty: Medium
 			other.crusher_loot = list(/obj/item/storm_staff, /obj/item/crusher_trophy/empowered_legion_skull)
 			return ..()
 	UnlockBlastDoors("11119")
-	..()
+	return ..()
 
-/mob/living/simple_animal/hostile/megafauna/legion/adjustHealth(damage, updating_health)
-	. = ..()
-	if(!GLOB.necropolis_gate)
-		return
-	if(GLOB.necropolis_gate.legion_triggered)
-		return
-	GLOB.necropolis_gate.toggle_the_gate(src, TRUE)
 
 /mob/living/simple_animal/hostile/megafauna/legion/AttackingTarget()
 	. = ..()
@@ -209,24 +206,33 @@ Difficulty: Medium
 				var/armor = M.run_armor_check(limb_to_hit, LASER)
 				M.apply_damage(70 - ((health / maxHealth) * 20), BURN, limb_to_hit, armor)
 
+
 /mob/living/simple_animal/hostile/megafauna/legion/Process_Spacemove(movement_dir = NONE)
 	return TRUE
 
+
 /mob/living/simple_animal/hostile/megafauna/legion/adjustHealth(amount, updating_health = TRUE)
 	. = ..()
-	if(QDELETED(src))
-		return
-	if(.)
-		update_transform((enraged ? 0.33 : 1) + (health / maxHealth))
-		if(amount > 0 && (enraged || prob(33)))
-			var/mob/living/simple_animal/hostile/asteroid/hivelordbrood/A
-			if(enraged)
-				A = new /mob/living/simple_animal/hostile/asteroid/hivelordbrood/legion/advanced(loc)
-			else
-				A = new /mob/living/simple_animal/hostile/asteroid/hivelordbrood/legion(loc)
-			A.GiveTarget(target)
-			A.friends = friends
-			A.faction = faction
+
+	if(GLOB.necropolis_gate && !GLOB.necropolis_gate.legion_triggered)
+		GLOB.necropolis_gate.toggle_the_gate(src, TRUE)
+
+	if(!. || QDELETED(src))
+		return .
+
+	// we shrink sprite until scaling reaches value of 0.5, no megafauna midgets plz
+	update_transform((0.5 + (health / maxHealth)) / current_size)
+
+	if(amount > 0 && (enraged || prob(33)))
+		var/mob/living/simple_animal/hostile/asteroid/hivelordbrood/A
+		if(enraged)
+			A = new /mob/living/simple_animal/hostile/asteroid/hivelordbrood/legion/advanced(loc)
+		else
+			A = new /mob/living/simple_animal/hostile/asteroid/hivelordbrood/legion(loc)
+		A.GiveTarget(target)
+		A.friends = friends
+		A.faction = faction
+
 
 /obj/item/gps/internal/legion
 	icon_state = null
